Added scoring and fail counting
Score increments on brick breaking
Fail count increments on failure
Displays fail message if 3 or more failures
Displays win message if 25 or more score count
