+++ Library/EditorInstance.json
@@ -1,0 +1,6 @@
+{
+	"process_id" : 22424,
+	"version" : "2018.3.7f1",
+	"app_path" : "C:/Program Files/Unity/Editor/Unity.exe",
+	"app_contents_path" : "C:/Program Files/Unity/Editor/Data"
+}
