+++ Library/StateCache/Hierarchy/3f9f35-96b0f7b4bb87ae545b.json
@@ -1,1 +1,1 @@
-﻿{"m_ExpandedPrefabGameObjectFileIDs":[],"m_ExpandedSceneGameObjectInstanceIDs":[],"m_ScrollY":0.0,"m_LastClickedFileID":299347988629779725,"m_LastClickedInstanceID":0}
+﻿{"m_ExpandedPrefabGameObjectFileIDs":[],"m_ExpandedSceneGameObjectInstanceIDs":[],"m_ScrollY":0.0,"m_LastClickedFileID":0,"m_LastClickedInstanceID":0}
