+++ Library/StateCache/Hierarchy/3f9f35-mainStage.json
@@ -1,1 +1,1 @@
-﻿{"m_ExpandedPrefabGameObjectFileIDs":[],"m_ExpandedSceneGameObjectInstanceIDs":[],"m_ScrollY":0.0,"m_LastClickedFileID":0,"m_LastClickedInstanceID":0}
+﻿{"m_ExpandedPrefabGameObjectFileIDs":[],"m_ExpandedSceneGameObjectInstanceIDs":[-1068,3156],"m_ScrollY":33.0,"m_LastClickedFileID":0,"m_LastClickedInstanceID":3082}
