+++ Library/StateCache/SceneView/78bf9a-96b0f7b4bb87ae545b.json
@@ -1,1 +1,1 @@
-﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":false,"audioPlay":false,"sceneViewState":{"showFog":false,"showMaterialUpdate":false,"showSkybox":false,"showFlares":false,"showImageEffects":false,"showParticleSystems":true},"in2DMode":false,"pivot":{"x":-1.2266566753387452,"y":2.4100000858306886,"z":0.0},"rotation":{"x":0.23423969745635987,"y":0.4491518437862396,"z":-0.12329064309597016,"w":0.8533427715301514},"size":10.957261085510254,"orthographic":false}
+﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":false,"audioPlay":false,"sceneViewState":{"showFog":false,"showMaterialUpdate":false,"showSkybox":false,"showFlares":false,"showImageEffects":false,"showParticleSystems":true},"in2DMode":false,"pivot":{"x":-1.3909144401550294,"y":2.2413253784179689,"z":0.05677080154418945},"rotation":{"x":0.24285832047462464,"y":0.44162753224372866,"z":-0.1255100518465042,"w":0.854536235332489},"size":534.2919921875,"orthographic":false}
