+++ Library/StateCache/SceneView/78bf9a-mainStage.json
@@ -1,1 +1,1 @@
-﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"showFog":false,"showMaterialUpdate":false,"showSkybox":false,"showFlares":false,"showImageEffects":false,"showParticleSystems":true},"in2DMode":false,"pivot":{"x":-1.5178567171096802,"y":-1.4950371980667115,"z":3.4164609909057619},"rotation":{"x":0.3418596386909485,"y":-0.25766023993492129,"z":0.09804333746433258,"w":0.8984045386314392},"size":5.309758186340332,"orthographic":false}
+﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"showFog":false,"showMaterialUpdate":false,"showSkybox":false,"showFlares":false,"showImageEffects":false,"showParticleSystems":true},"in2DMode":false,"pivot":{"x":-1.5178567171096802,"y":-1.4950371980667115,"z":3.4164609909057619},"rotation":{"x":0.3418596386909485,"y":-0.25766023993492129,"z":0.09804333746433258,"w":0.8984045386314392},"size":29.095645904541017,"orthographic":false}
